Mouse click and symbol generation in UI completed.
Switched to using event.mouseButton to detect mouse clicks and using the coordinates of the position of the mouse when clicked using the event.mouseButton.x and y to know the where it was clicked.

Completed switching to using an array that saves the squares already played. When a player clicks on a square the array receives an update of its value to X or O in the position played.

The array with the coordinates played is then provided to the generateMap() function to draw the board, crosses and circles in the played locations.

To draw the crosses and squares with appropriate shape and in the correct locations, another array was created locally in the generateMap function which has the coordinates of the centers of each square. These coordinates are then provided to the generateCircles and generateCrosses functions to draw them in the correct locations.