Corrected game breaking beahaviour where the game crashed when a player clicked outside the squares
Include check to see if the location click is valid to be played. If it is, it proceeds, and if not it does nothing.

Added function generateTitles to print any text to the window. So far prints the name of the game "Tic Tac Toe" over the board.

Started analysing views to handle game window size changes.