Created roundEnded function to process who was the player that won the game or if it was a draw.
Optimized the event processing for left mouse click to just save the mouse location when clicked in the leftMouseClickCoords int array.

Move the code to change the symbol in the board array to outside the event procesing code.

Created the generatePlayAgainBox function to draw a pop up box after a player wins the game or ends in a draw. Full working implementation still not done.